requirements.txt
PySide6
Pillow
sqlalchemy
psycopg2-binary
pgvector
python-dotenv